+++ app/components/UI/Rewards/components/Campaigns/CampaignOptInCta.tsx
@@ -1,15 +1,18 @@
-import React, { useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import {
   Box,
   Button,
   ButtonVariant,
   ButtonSize,
+  IconName,
 } from '@metamask/design-system-react-native';
 import type { CampaignDto } from '../../../../../core/Engine/controllers/rewards-controller/types';
 import type { UseGetCampaignParticipantStatusResult } from '../../hooks/useGetCampaignParticipantStatus';
 import CampaignOptInSheet from './CampaignOptInSheet';
 import { getCampaignStatus } from './CampaignTile.utils';
 import { strings } from '../../../../../../locales/i18n';
+import useCampaignGeoRestriction from '../../hooks/useCampaignGeoRestriction';
+import useRewardsToast from '../../hooks/useRewardsToast';
 
 export const CAMPAIGN_CTA_TEST_IDS = {
   CTA_BUTTON: 'campaign-details-cta-button',
@@ -22,19 +25,28 @@ interface CampaignOptInCtaProps {
     'status' | 'isLoading'
   >;
   onJoinPress?: () => void;
+  /** An optional set of country codes that are restricted independently of the campaign's `excludedRegions`. Checked before `excludedRegions`. When the user's country matches this list the CTA shows "Check eligibility" and a geo-locked toast instead of opening the opt-in sheet. */
+  customRestrictedCountries?: Set<string>;
 }
 
 /**
  * General campaign opt-in CTA.
- * Shows a "Join Campaign" button when the campaign is active, the user has not opted in,
- * and the opt-in window is still open. Returns null in all other cases.
+ * When the user is geo-restricted: shows an enabled "Check eligibility" button (lock icon) that fires a "not available in your region" toast.
+ * Otherwise: shows a "Join Campaign" button that opens the opt-in sheet.
+ * Returns null when the campaign is not active or the user is already opted in.
  */
 const CampaignOptInCta: React.FC<CampaignOptInCtaProps> = ({
   campaign,
   participantStatus,
   onJoinPress,
+  customRestrictedCountries,
 }) => {
   const [isOptInSheetOpen, setIsOptInSheetOpen] = useState(false);
+  const { showToast, RewardsToastOptions } = useRewardsToast();
+  const { isGeoRestricted } = useCampaignGeoRestriction(
+    campaign,
+    customRestrictedCountries,
+  );
 
   const campaignStatus = getCampaignStatus(campaign);
   const isLoading = participantStatus.isLoading;
@@ -42,8 +54,34 @@ const CampaignOptInCta: React.FC<CampaignOptInCtaProps> = ({
 
   const isActive = !isLoading && campaignStatus === 'active';
 
+  const handleGeoLockedPress = useCallback(() => {
+    showToast(
+      RewardsToastOptions.entriesClosed(
+        strings('rewards.campaign_details.ondo.geo_locked_toast_title'),
+        strings('rewards.campaign_details.ondo.geo_locked_toast_description'),
+      ),
+    );
+  }, [showToast, RewardsToastOptions]);
+
   if (!isActive || isOptedIn) {
     return null;
+  }
+
+  if (isGeoRestricted) {
+    return (
+      <Box twClassName="p-4 mb-2">
+        <Button
+          variant={ButtonVariant.Primary}
+          size={ButtonSize.Lg}
+          isFullWidth
+          startIconName={IconName.Lock}
+          onPress={handleGeoLockedPress}
+          testID={CAMPAIGN_CTA_TEST_IDS.CTA_BUTTON}
+        >
+          {strings('rewards.campaign_details.ondo.geo_locked_cta')}
+        </Button>
+      </Box>
+    );
   }
 
   return (
@@ -67,6 +105,7 @@ const CampaignOptInCta: React.FC<CampaignOptInCtaProps> = ({
         <CampaignOptInSheet
           campaign={campaign}
           onClose={() => setIsOptInSheetOpen(false)}
+          customRestrictedCountries={customRestrictedCountries}
         />
       )}
     </>
