+++ app/components/UI/Rewards/components/Campaigns/OndoCampaignCTA.tsx
@@ -17,6 +17,7 @@ import CampaignOptInCta, { CAMPAIGN_CTA_TEST_IDS } from './CampaignOptInCta';
 import OndoNotEligibleSheet from './OndoNotEligibleSheet';
 import { useAnalytics } from '../../../../hooks/useAnalytics/useAnalytics';
 import { MetaMetricsEvents } from '../../../../../core/Analytics';
+import { ONDO_RESTRICTED_COUNTRIES } from '../../../../../util/ondoGeoRestrictions';
 
 interface OndoCampaignCTAProps {
   campaign: CampaignDto;
@@ -27,13 +28,12 @@ interface OndoCampaignCTAProps {
   hasPositions: boolean;
   campaignId: string;
   notEligibleForCampaign?: boolean;
-  isGeoRestricted?: boolean;
 }
 
 /**
  * Bottom CTA for the Ondo campaign details page.
  * Renders one of four states depending on campaign/participant status:
- * - Delegates to CampaignCTA for the opt-in flow (active, not opted in, within deposit window)
+ * - Delegates to CampaignOptInCta for the opt-in flow (active, not opted in, within deposit window). Passes ONDO_RESTRICTED_COUNTRIES so that CampaignOptInCta shows the geo-locked "Check eligibility" CTA for restricted users.
  * - "Entries closed" button (with Lock icon + toast) when cutoff has passed and user is not opted in
  * - "Open Position" button when the user has opted in but has no portfolio positions
  * - "Swap Ondo Assets" button when the user has opted in and has portfolio positions
@@ -44,7 +44,6 @@ const OndoCampaignCTA: React.FC<OndoCampaignCTAProps> = ({
   hasPositions,
   campaignId,
   notEligibleForCampaign = false,
-  isGeoRestricted = false,
 }) => {
   const navigation = useNavigation();
   const { trackEvent, createEventBuilder } = useAnalytics();
@@ -119,15 +118,6 @@ const OndoCampaignCTA: React.FC<OndoCampaignCTAProps> = ({
     );
   }, [showToast, RewardsToastOptions]);
 
-  const handleGeoLockedPress = useCallback(() => {
-    showToast(
-      RewardsToastOptions.entriesClosed(
-        strings('rewards.campaign_details.ondo.geo_locked_toast_title'),
-        strings('rewards.campaign_details.ondo.geo_locked_toast_description'),
-      ),
-    );
-  }, [showToast, RewardsToastOptions]);
-
   if (isEntriesClosed) {
     return (
       <Box twClassName="px-4 pt-2">
@@ -150,23 +140,6 @@ const OndoCampaignCTA: React.FC<OndoCampaignCTAProps> = ({
     return null;
   }
 
-  if (!isOptedIn && isGeoRestricted) {
-    return (
-      <Box twClassName="px-4 pt-2">
-        <Button
-          variant={ButtonVariant.Primary}
-          size={ButtonSize.Lg}
-          isFullWidth
-          startIconName={IconName.Lock}
-          onPress={handleGeoLockedPress}
-          testID={CAMPAIGN_CTA_TEST_IDS.CTA_BUTTON}
-        >
-          {strings('rewards.campaign_details.ondo.geo_locked_cta')}
-        </Button>
-      </Box>
-    );
-  }
-
   if (!isOptedIn) {
     if (notEligibleForCampaign) {
       return (
@@ -187,6 +160,7 @@ const OndoCampaignCTA: React.FC<OndoCampaignCTAProps> = ({
       <CampaignOptInCta
         campaign={campaign}
         participantStatus={participantStatus}
+        customRestrictedCountries={ONDO_RESTRICTED_COUNTRIES}
         onJoinPress={() =>
           trackEvent(
             createEventBuilder(MetaMetricsEvents.REWARDS_PAGE_BUTTON_CLICKED)
